+++ Changelog.md
@@ -1,9 +1,10 @@
 # Changelog
 
 ## 0.5.2 (Nov 2023)
-* `pyaudio` is now optional: If you plan to use `PyAudioBackend`, install `pya` with `pip install pya[pyaudio]`
-* Fix audio device bug 
+* #82, `pyaudio` is now optional: If you plan to use `PyAudioBackend`, install `pya` with `pip install pya[pyaudio]`.
+* Fix audio device bug
 * #77, Improve code type hint
+* #79, Use pyamapping 
 
 ## 0.5.1 (Dec 2022)
 * Now support Python3.10